+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for the `#include` directive
 - `sw-frag` tool to get the complete combined fragment shader
 - Play/Pause/Reset controls
+- Canvas screenshot
 
 ### Changed
 - `shader-workshop` program has been renamed to `sw-server`